requirements.txt
absl-py==2.3.0
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.4
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
evaluate==0.4.3
exceptiongroup==1.3.0
executing==2.2.0
ffmpeg-python==0.2.0
filelock==3.18.0
flash-attn==2.7.4.post1
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
gitpython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
ipykernel==6.29.5
ipython==8.36.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.8
markdown==3.8
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==0.28.1
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.9
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
qwen-vl-utils==0.0.11
regex==2024.11.6
replicate==1.0.7
requests==2.32.3
rouge-score==0.1.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchcodec==0.4.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
triton==3.2.0
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
wandb==0.19.11
wcwidth==0.2.13
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
